requirements.txt
Flask==2.2.2
PyMuPDF==1.19.6
python-docx==0.8.11
requests==2.28.1
Werkzeug==2.2.2
flask-wtf==1.0.0
python-dotenv==0.20.0